+++ fitbit/exceptions.py
