+++ fitbit/utils.py
