+++ src/lib/client.py
@@ -47,16 +47,3 @@ class Fitbit():
             }
 
         return self.apply_converter(convert, weights)
-
-    def get_activities(self, start_date, end_date):
-        activities = self.client.activities(date="2020-01-13")
-
-        def convert(data):
-            data
-            # return {
-            #     "date": weight["date"],
-            #     "weight": pound_to_kg(weight["weight"]),
-            #     "bmi": weight["bmi"]
-            # }
-
-        return self.apply_converter(convert, activities)
